Add ability for admin to add products with AWS
Admin can now use dashboard to add products, uploading images to AWS bucket.
Figaro is used to manage ENV vars and some js is used to make the file
uplaod button nice. If you inspect the upload element, it's actually hidden
and being swaped out for a nicer button with js since the styling for file
upload forms in html is notoriously difficult.

Tests are now running super slowly because everytime we create a product we
have to make a call to the AWS server. We should brainstorm how to fix this.
I sent a good article about this topic via slack to the team.