Platform Admin Approves Listing
Platform admin now can view and change the attributes for every pending
property that exists within the database from the users. This comes with a
search bar feature that allows a user to grab either a specific owner name,
property id, or date when property was listed and pending.

Bugs that need to be fixed:
- Headers for specific property view that is either approved or pending

closes#40