Platform Admin Disables Property
Our table for properties no longer has an "approved" column that is a boolean.
Instead, it is now a "status" column that is a string with a default value
"pending." This also fixed the bug that was happening within the Admin's
view when they'd select one of the tabs that had to do with viewing pending,
active, and now disabled properties.

closes#41